feat: Added module metadata and updated service methods
In this commit, a module metadata has been added to keep track of different modules in the application. This includes modules name, version, description, release state, and other details. Some changes have also been done in the service method in the ModuleSettingsService.cs file, especially related to querying the database and aggregating results. Further, some updates are done in the programs.cs within the Worker directory, registering the modules in the ConfigureServices. Lastly, certain sonarlint paths in the .pb files have been updated.